Long time no push because i forgot. Players can now collide with each other, pathfinding operations are in a different script and theres a non-bug-tested solution for collision if both players are moving. OpenGl drawing errors in colsole fixed

+++ Terrain.hpp
@@ -7,7 +7,7 @@ public:
 	Terrain();
 	void addRect(int row, int col, int width, int height);
 	void draw();
-	void addCollidablesToGrid(bool** grid, int pathfindingAccuracy, int playerWidth, int playerHeight);
+	void addCollidablesToGrid(bool** grid, float pathfindingAccuracy, int playerWidth, int playerHeight);
 	std::vector<Rect*>* getCollidables();
 
 private:
